[FAB2023-999] feat: 카테고리 추가/변경 모달 개발 - 컴포넌트 화, 공통 기능 store로 이동
