+++ project/ovp/frontend/ovp/components/govern/category/category-add-modal.vue
@@ -1,0 +1,85 @@
+<template>
+  <Modal
+    title="카테고리 추가"
+    class="modal modal-padding-16"
+    :modal-id="props.modalId"
+    background="non-interactive"
+    displayDirective="show"
+    overlayTransition="vfm-fade"
+    contentTransition="vfm-fade"
+    :clickToClose="true"
+    :escToClose="true"
+    :width="480"
+    :height="430"
+    :top="380"
+    :lockScroll="false"
+    swipeToClose="none"
+    @cancel="onCancel"
+    @confirm="onConfirm"
+  >
+    <template v-slot:body>
+      <div class="form form-lg">
+        <div class="form-body">
+          <div class="form-item">
+            <label for="data-model-save-name" class="form-label">
+              이름
+              <span class="required">*</span>
+            </label>
+            <div class="form-detail">
+              <input
+                id="data-model-save-name"
+                class="text-input text-input-lg"
+                placeholder="이름을 입력하세요."
+              />
+              <div class="notification notification-sm notification-error">
+                <svg-icon class="notification-icon" name="error"></svg-icon>
+                <p class="notification-detail">이름을 입력하세요.</p>
+              </div>
+            </div>
+          </div>
+          <div class="form-item">
+            <label class="form-label" for="data-model-save-description">
+              설명
+              <span class="required">*</span>
+            </label>
+            <div class="form-detail">
+              <textarea
+                id="data-model-save-description"
+                class="textarea h-28"
+                placeholder="설명을 입력하세요."
+              ></textarea>
+              <div class="notification notification-sm notification-error">
+                <svg-icon class="notification-icon" name="error"></svg-icon>
+                <p class="notification-detail">설명을 입력하세요.</p>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </template>
+  </Modal>
+</template>
+
+<script setup lang="ts">
+import Modal from "@extends/modal/Modal.vue";
+import { useNuxtApp } from "nuxt/app";
+
+const { $vfm } = useNuxtApp();
+
+const props = defineProps({
+  modalId: {
+    type: String,
+    required: true,
+  },
+});
+
+const onCancel = () => {
+  $vfm.close(props.modalId);
+};
+
+const onConfirm = async () => {
+  $vfm.open(props.modalId);
+};
+</script>
+
+<style scoped></style>
